+++ python/src/ouster/client/__init__.py
@@ -8,6 +8,7 @@ from ._client import OperatingMode
 from ._client import MultipurposeIOMode
 from ._client import Polarity
 from ._client import NMEABaudRate
+from ._client import ChanField
 from ._client import SensorConfig
 from ._client import get_config
 from ._client import set_config
@@ -17,7 +18,6 @@ from .data import BufferT
 from .data import Packet
 from .data import ImuPacket
 from .data import LidarPacket
-from .data import ChanField
 from .data import ColHeader
 from .data import XYZLut
 from .data import destagger